+++ README.md
@@ -1,10 +1,15 @@
 # MarkoMeter
 ##   MarkoMeter	: 	Student Result Management System
 ##   Author 	: 	Vedant Khandelwal
-##   Version 	: 	2.0
+##   Version 	: 	1.0
 ##   Github 	: 	https://github.com/TheVedantKhandelwal
-  
+ 
+ # Installation - 
+ Launch the command line and copy paste the lines in quotes
+ "pip install matplotlib,pandas,numpy"
+ 
+ 
  ## Setup : -
     Create a CSV with columns - 'Physics','Chemistry','Maths' in the Documents Folder
-    Open  in a text editor and Search for "YOUR PATH HERE" in every '.py' file, replace text in quotes with the path to the CSV file you just made
-    Run main.py :)
+    Open 'markometer.py' in a text editor and Search for "YOUR PATH HERE", replace text in quotes with the path to the CSV file you just made
+
